FizzBuzz: Move TextDisplay to a separate file
Extract the display to its own file to separate functionality from testing behavior
