+++ clean-code-katas/function-katas/fizzbuzz/TextDisplay.cs
@@ -1,0 +1,15 @@
+﻿using System;
+using System.Collections.Generic;
+
+namespace fizzbuzz
+{
+    internal class TextDisplay
+    {
+        public void DisplayTexts(List<string> texts, Action<string> displayAction)
+        {
+            var output = string.Join(",", texts);
+
+            displayAction(output);
+        }
+    }
+}
