+++ clean-code-katas/function-katas/fizzbuzz/fizzbuzz.csproj
@@ -44,6 +44,7 @@
     <Compile Include="FizzBuzzConverter.cs" />
     <Compile Include="NumberGenerator.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="TextDisplay.cs" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
 </Project>
